Fixed return data types in methods

+++ StRoot/StEpdDbMaker/StEpdDbMaker.h
@@ -44,10 +44,10 @@ class StEpdDbMaker : public StMaker{
     short GetStatus(short ew, short pp, short tile){return mStatus[ew][pp][tile];}
 
     // epdGain table
-    short GetVPed(short ew, short pp, short tile){return mVPed[ew][pp][tile];}
-    short GetMip(short ew, short pp, short tile){return mMip[ew][pp][tile];}
-    short GetQtPedestals(short ew, short pp, short tile){return mQtPedestals[ew][pp][tile];}
-    short GetDarkCurrent(short ew, short pp, short tile){return mDarkCurrent[ew][pp][tile];}
+    float GetVPed(short ew, short pp, short tile){return mVPed[ew][pp][tile];}
+    float GetMip(short ew, short pp, short tile){return mMip[ew][pp][tile];}
+    float GetQtPedestals(short ew, short pp, short tile){return mQtPedestals[ew][pp][tile];}
+    float GetDarkCurrent(short ew, short pp, short tile){return mDarkCurrent[ew][pp][tile];}
 
   protected:
